# api/openapi.yaml
openapi: 3.0.3
info:
  title: Aether API
  description: API for managing AI agents with Firecracker isolation
  version: 0.2.0
  contact:
    name: Aether Development Team
servers:
  - url: http://localhost:8080/v1
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      description: Check if the API is running
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  time:
                    type: string
                    format: date-time

  /agents:
    get:
      summary: List agents
      description: List all agents (filtered by tenant)
      operationId: listAgents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentInfo'

    post:
      summary: Create agent
      description: Create and start a new agent
      operationId: createAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentConfig'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/QuotaExceeded'

  /agents/{id}:
    get:
      summary: Get agent
      description: Get information about a specific agent
      operationId: getAgent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentInfo'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete agent
      description: Stop and remove an agent
      operationId: deleteAgent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Agent deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /agents/{id}/health:
    get:
      summary: Check agent health
      description: Get health status of an agent
      operationId: getAgentHealth
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /quotas:
    get:
      summary: List quotas
      description: List all tenant quotas
      operationId: listQuotas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of quotas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quota'

  /quotas/{tenant_id}:
    get:
      summary: Get quota
      description: Get quota for a specific tenant
      operationId: getQuota
      security:
        - bearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant quota
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quota'

    put:
      summary: Set quota
      description: Set or update quota for a tenant
      operationId: setQuota
      security:
        - bearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Quota'
      responses:
        '200':
          description: Quota updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quota'

  /scheduler/stats:
    get:
      summary: Get scheduler statistics
      description: Get current scheduler statistics
      operationId: getSchedulerStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scheduler statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerStats'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AgentConfig:
      type: object
      required:
        - name
        - image
      properties:
        id:
          type: string
          description: Unique agent ID (auto-generated if not provided)
        tenant_id:
          type: string
          description: Tenant ID (extracted from auth if not provided)
        name:
          type: string
          description: Human-readable agent name
        image:
          type: string
          description: Container image to run
        resources:
          $ref: '#/components/schemas/ResourceLimits'
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
        secrets:
          type: array
          items:
            type: string
          description: Secret keys to inject

    AgentInfo:
      type: object
      properties:
        config:
          $ref: '#/components/schemas/AgentConfig'
        status:
          type: string
          enum: [pending, creating, running, stopping, stopped, failed]
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        stopped_at:
          type: string
          format: date-time
        error:
          type: string
        metrics:
          $ref: '#/components/schemas/AgentMetrics'

    ResourceLimits:
      type: object
      properties:
        cpu_count:
          type: integer
          minimum: 1
          default: 1
        memory_mb:
          type: integer
          minimum: 128
          default: 512
        disk_mb:
          type: integer
          minimum: 0

    AgentMetrics:
      type: object
      properties:
        cpu_usage_percent:
          type: number
          format: float
        memory_usage_mb:
          type: integer
        network_rx_bytes:
          type: integer
        network_tx_bytes:
          type: integer
        last_updated:
          type: string
          format: date-time

    HealthStatus:
      type: object
      properties:
        healthy:
          type: boolean
        message:
          type: string
        checked_at:
          type: string
          format: date-time

    Quota:
      type: object
      properties:
        tenant_id:
          type: string
        max_agents:
          type: integer
        max_cpu_cores:
          type: integer
          description: Maximum CPU cores in millicores
        max_memory_mb:
          type: integer
        max_disk_mb:
          type: integer
        max_requests_per_minute:
          type: integer
        tier:
          type: string
          enum: [free, pro, enterprise]

    SchedulerStats:
      type: object
      properties:
        node_count:
          type: integer
        agent_count:
          type: integer
        queued_requests:
          type: integer
        total_capacity:
          $ref: '#/components/schemas/Resources'
        total_allocated:
          $ref: '#/components/schemas/Resources'

    Resources:
      type: object
      properties:
        cpu_cores:
          type: integer
        memory_mb:
          type: integer
        disk_mb:
          type: integer

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string

    QuotaExceeded:
      description: Quota limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string